+++ resources/views/pages/wilayah/wilayah.blade.php
@@ -294,27 +294,28 @@
             let url = $(this).attr('data-url');
             let id = $(this).attr('data-id');
             var token = $("meta[name='csrf-token']").attr("content");
-
-            $.ajax({
-                url: url,
-                type: 'DELETE',
-                data: {
-                    "id": id,
-                    "_token": token,
-                },
-                success: function(res) {
-                    if (res) {
-                        Toastify({
-                            text: 'Data Berhasil Dihapus',
-                            className: "success",
-                            style: {
-                                background: "#00b09b",
-                            }
-                        }).showToast();
+            if (confirm("Yakin Data Di Hapus ?")) {
+                $.ajax({
+                    url: url,
+                    type: 'DELETE',
+                    data: {
+                        "id": id,
+                        "_token": token,
+                    },
+                    success: function(res) {
+                        if (res) {
+                            Toastify({
+                                text: 'Data Berhasil Dihapus',
+                                className: "success",
+                                style: {
+                                    background: "#00b09b",
+                                }
+                            }).showToast();
+                        }
+                        location.reload();
                     }
-                    location.reload();
-                }
-            });
+                });
+            }
         });
     </script>
